refactor : getKoreaTodayDateInfo,getWeekDateLsit util 함수 작성
